Update padding in rows and card body
